requirements.txt
# Core Dependencies
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.22
openai==1.86.0
python-dotenv==1.1.0

# Document Processing
python-docx==1.1.2
pypdf==5.6.0
pandas==2.3.0
docling==2.36.1
docling-core==2.37.0

# Vector Storage
chromadb==1.0.12
tiktoken==0.9.0

# Web UI
streamlit==1.45.1
fastapi==0.115.9
uvicorn==0.34.3

# Evaluation & Metrics
scikit-learn>=1.0.0
rouge-score>=0.1.2
nltk>=3.6.0

# Development Tools
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
python-dotenv==1.1.0

# Utilities
tqdm==4.67.1
numpy==2.2.6
pydantic==2.11.5
